Created login page that returns success/fail
